% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{brandthis_ragnar_store}
\alias{brandthis_ragnar_store}
\title{Open the ggsci knowledge store}
\usage{
brandthis_ragnar_store(pkg = c("paletteer", "ggsci"))
}
\arguments{
\item{pkg}{ggsci or paletteer}
}
\value{
A \code{RagnarStore} object.
}
\description{
Opens a connection to the ggsci documentation database. On first use,
this will download the database from GitHub.
}
\keyword{internal}
