% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand.R
\name{chat_brand}
\alias{chat_brand}
\title{Create a chat object for generating _brand.yml}
\usage{
chat_brand(chat_fn = ellmer::chat_google_gemini, ...)
}
\arguments{
\item{chat_fn}{Ellmer chat function, e.g., \code{ellmer::chat_google_gemini}.}

\item{...}{Parameters of ellmer chat function.}
}
\value{
An ellmer chat object
}
\description{
Create a chat object for generating _brand.yml
}
