% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand.R
\name{create_brand}
\alias{create_brand}
\title{Create _brand.yml}
\usage{
create_brand(
  prompt,
  img = NULL,
  browse = FALSE,
  type = c("personal", "company"),
  chat_fn = ellmer::chat_google_gemini,
  ...
)
}
\arguments{
\item{prompt}{User prompt. May include info. about brand name, colors, fonts, and logo.}

\item{img}{Image URL(s) and/or path(s). Only path(s) are allowed for \code{type = "company"}.}

\item{browse}{Whether to browse for image(s) in the file system. Default is \code{FALSE}.}

\item{type}{Brand type. \code{"personal"} or \code{"company"}.}

\item{chat_fn}{Ellmer chat function, e.g., \code{ellmer::chat_google_gemini}.}

\item{...}{Parameters of ellmer chat function.}
}
\value{
_brand.yml object
}
\description{
Create _brand.yml
}
