% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_color_palettes.R
\name{create_color_palette}
\alias{create_color_palette}
\title{Create color palettes from the provided _brand.yml}
\usage{
create_color_palette(brand_yml, chat_fn = ellmer::chat_google_gemini, ...)
}
\arguments{
\item{brand_yml}{_brand.yml object.}

\item{chat_fn}{Ellmer chat function, e.g., \code{ellmer::chat_google_gemini}.}

\item{...}{Parameters of ellmer chat function.}
}
\value{
List of color palettes
}
\description{
Create color palettes from the provided _brand.yml
}
\examples{
\dontrun{
personal_brand <- create_brand(
  "My name is Paul Blart",
  img = "https://free-images.com/md/7687/blue_jay_bird_nature.jpg",
  type = "personal",
  chat_fn = ellmer::chat_google_gemini
)
color_palettes <- create_color_palette(personal_brand)
}
}
