% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_palette_from_image}
\alias{create_palette_from_image}
\title{Create a color palette from an image}
\usage{
create_palette_from_image(
  img = NULL,
  n = 5,
  extract_method = "median",
  show = FALSE,
  ...
)
}
\arguments{
\item{img}{Image URL(s) and/or path(s).}

\item{n}{Number of colors. Default is 5.}

\item{extract_method}{Method of color extraction. Default is the \code{median} of the cluster specific RGB values. Other options are \code{"mean"}, \code{"mode"}, and \code{"hex_freq"} (most common hex colors per cluster)}

\item{show}{Whether to show the generated color palette. Default is \code{FALSE}}

\item{...}{Optional arguments in \code{colorfindr::get_colors}}
}
\value{
Vector of colors
create_palette_from_image("https://free-images.com/lg/871f/garden_violet_blooms_profusely_0.jpg")
}
\description{
Create a color palette from an image
}
