% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fonts_combination}
\alias{get_fonts_combination}
\title{Get a combination of heading and base Google Fonts}
\usage{
get_fonts_combination()
}
\value{
Character vector with a single element containing a fonts combination
}
\description{
Get a combination of heading and base Google Fonts
}
