% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_image_paths}
\alias{get_image_paths}
\title{Get image URL(s) and/or path(s)}
\usage{
get_image_paths(img = NULL, browse = FALSE, type = c("personal", "company"))
}
\arguments{
\item{img}{Image URL(s) and/or path(s).}

\item{browse}{Whether to browse for image(s) in the file system. Default is \code{FALSE}.}

\item{type}{Brand type. \code{"personal"} or \code{"company"}.}
}
\value{
Image URL(s) and/or path(s) as a character vector.
}
\description{
Get image URL(s) and/or path(s)
}
