% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{semantic_colors_as_hex_codes}
\alias{semantic_colors_as_hex_codes}
\title{Get hex codes of all semantic colors}
\usage{
semantic_colors_as_hex_codes(color_list)
}
\arguments{
\item{color_list}{List containing \code{palette} and semantic colors}
}
\value{
List of hex codes for each semantic color
}
\description{
Get hex codes of all semantic colors
}
