% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_color_scales.R
\name{suggest_color_scales}
\alias{suggest_color_scales}
\title{Suggest ggsci color palettes and scales}
\usage{
suggest_color_scales(brand_yml, chat_fn = ellmer::chat_google_gemini, ...)
}
\arguments{
\item{brand_yml}{_brand.yml object}

\item{chat_fn}{Ellmer chat function, e.g., \code{ellmer::chat_google_gemini}.}

\item{...}{Parameters of ellmer chat function.}
}
\value{
Suggested palettes and scales
}
\description{
Suggest ggsci color palettes and scales
}
