% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store.R
\name{update_store}
\alias{update_store}
\title{Updates the ggsci knowledge store}
\usage{
update_store(pkg = c("paletteer", "ggsci"))
}
\arguments{
\item{pkg}{ggsci or paletteer}
}
\value{
\code{NULL} invisibly.
}
\description{
Downloads the latest version of the store from the brandthis GitHub
repository and builds the search index locally.
}
\details{
The download location can be configured with a few environment variables:
\itemize{
\item \code{BRANDTHIS_STORE_URL}: a custom URL to download the store from. The default is to download the latest store
from the \code{durraniu/brandthis} repository in the \code{store-v1} release.
\item \code{BRANDTHIS_STORE_RELEASE}: the release tag to download the store from. Defaults to \code{store-v1}.
\item \code{BRANDTHIS_STORE_REPOSITORY}: the repository to download the store from. Defaults to \code{durraniu/brandthis}.
}
}
