% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brand.R
\name{create_brand}
\alias{create_brand}
\title{Create _brand.yml}
\usage{
create_brand(
  prompt,
  img = NULL,
  browse = FALSE,
  type = c("personal", "company"),
  chat_fn = ellmer::chat_google_gemini,
  ...
)
}
\arguments{
\item{prompt}{User prompt. May include info. about brand name, colors, fonts, and logo.}

\item{img}{Image URL(s) and/or path(s). Only path(s) are allowed for \code{type = "company"}. Default is \code{NULL}.}

\item{browse}{Whether to browse for image(s) in the file system. Default is \code{FALSE}.}

\item{type}{Brand type. \code{"personal"} or \code{"company"}. See Details.}

\item{chat_fn}{\href{https://ellmer.tidyverse.org/#providers}{Ellmer} chat function. Default is \code{ellmer::chat_google_gemini}.}

\item{...}{Parameters of ellmer chat function.}
}
\value{
_brand.yml object
}
\description{
Create a personal/company \code{_brand.yml} file using your favourite LLM.
}
\details{
Generating _brand.yml requires info. about your brand name, colors, typography,
and logos. If no info. is provided, \code{create_brand} will ask for it.
Brand name and image(s) or colors are required at a minimum.
\code{create_brand} will use a random combination of Google Fonts if no
typography info. is provided.

Personal Branding: You may provide a single image path or a URL that \code{create_brand} will use
to extract colors as the first step. These colors will then be used by an LLM
to generate the color palette and semantic colors for the _brand.yml.
If an image is not provided, the LLM will use info. from the \code{prompt}.

Company Branding: You may provide one or more image paths of company branding guidelines that
an LLM will extract info. from for creating the _brand.yml. Image URLs are
not allowed. If no image(s) are provided, \code{create_brand} will rely on
the \code{prompt}.
}
\examples{
\dontrun{
# Create a personal _brand.yml with a local image
personal_brand <- create_brand(
  "My name is Katniss Everdeen",
  browse = TRUE,
  type = "personal",
  chat_fn = ellmer::chat_github,
  model = "gpt-4.1"
)

# Create a personal _brand.yml with an image URL
personal_brand <- create_brand(
  "My name is Paul Blart",
  img = "https://free-images.com/md/7687/blue_jay_bird_nature.jpg",
  type = "personal",
  chat_fn = ellmer::chat_google_gemini
)

# If no info. is provided, you will be asked to provide some:
brand_no_input <- create_brand(
  "",
  img = NULL,
  type = "personal",
  chat_fn = ellmer::chat_google_gemini
)
# I can help with that! To create your personal
# `_brand.yml` file, I'll need a bit more
# information. Please provide the following:
#
# *   **Brand Name** (e.g., John Doe, or a short/full
# name) and **Website(s)** (e.g., personal portfolio,
# social media links)
# *   **Brand Colors** or a **Color Palette** (e.g.,
# "I like blue, green, and a touch of orange," or
# specific hex codes)
# *   **Heading, Base, and Monospace Fonts** (e.g.,
# "I'd like Roboto for headings, Open Sans for base
# text, and Fira Code for monospace," or you can ask
# me to suggest some)
# *   **Logo information** (e.g., "I have a logo at
# this URL," or "I'd like a placeholder for now")

# Create a company _brand.yml with multiple image paths:
company_brand <- create_brand(
  "Company name is Walmart",
  img = c("walmart-font.png",
          "walmart-palette.jpeg",
          "walmart-logo.png"),
  type = "company",
  chat_fn = ellmer::chat_google_gemini
)
}
}
