% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_color_scales.R
\name{suggest_color_scales}
\alias{suggest_color_scales}
\title{Suggest paletteer or ggsci color palettes and scales}
\usage{
suggest_color_scales(
  brand_yml,
  pkg = c("paletteer", "ggsci"),
  top_k = 8L,
  chat_fn = ellmer::chat_google_gemini,
  ...
)
}
\arguments{
\item{brand_yml}{_brand.yml object.}

\item{pkg}{paletteer or ggsci}

\item{top_k}{Number of excerpts to request from the knowledge store for each retrieval}

\item{chat_fn}{Ellmer chat function, e.g., \code{ellmer::chat_google_gemini}.}

\item{...}{Parameters of ellmer chat function.}
}
\value{
Suggested palettes and scales
}
\description{
Suggest paletteer or ggsci color palettes and scales
}
\examples{
\dontrun{
personal_brand <- create_brand(
  "My name is Paul Blart",
  img = "https://free-images.com/md/7687/blue_jay_bird_nature.jpg",
  type = "personal",
  chat_fn = ellmer::chat_google_gemini
)
suggested_scales <- suggest_color_scales(personal_brand)
}
}
